dry shared cli functions
